+++ pants-plugins/oci/pants_backend_oci/goals/__init__.py
@@ -1,4 +1,4 @@
-from pants_backend_oci.goals import package, package_build_step, publish, run
+from pants_backend_oci.goals import package, package_build_step, publish, run, test
 
 
 def rules():
@@ -7,4 +7,5 @@ def rules():
         *package.rules(),
         *package_build_step.rules(),
         *run.rules(),
+        *test.rules(),
     ]