+++ pants-plugins/oci/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Add new `oci_structure_test` target. As a start, it allows validating the digest of the image.
+
 ## 0.4.0 - 2023-06-18
 
 - Improved support for very large layers > 2GB. A lot of layers will now be compressed in